Fix first-time issue with generated storyLoader.js
This adds the generated file to the repo rather than ignoring it.
While this requires us to commit it every time we add a new story,
it avoids the problem where `yarn start` fails to import it if you
haven't first run `yarn storybook`.

There may be more elegant solutions to this problem, which we could
consider in the future.