Style drawer content
Since our drawer actions are actually all modal screens, and since
that's not the best supported use of react-navigation's DrawerNavigator,
I've switched to using a DrawerLayout from react-native-gesture-handler.

https://github.com/react-navigation/react-navigation/issues/3447#issuecomment-364497083

Bonus: this also lets us do the styling we want, without waiting
for react-navigation 3.0. We're also slightly more future-proof
in the event we want some of the drawer items to not be modals.