Gracefully handle Patreon device verification
The first time a user logs into Patreon from a new device, it goes
through an email-based device verification. The user has to open their
email and click a link, then try logging in again. Since this breaks the
oauth flow in our app, we simply navigate back to the patreon screen and
display a message reminding the user to check their email and try again.

Other Patreon fixes:
- Added text distinguishing between not-connected and not-a-patron
- Updated patreon campaign URL (it was changed in Patreon at some point)